[EE] Prettify compiler-generated field names.
Issue:
- The debugger's NullReferenceException analysis feature figures out what was null through metadata and calls into the EE to display the offender as an expression.
- If a field reference was null, we end up calling into GetClrMemberName.
- However, for hoisted locals, etc we end up with an ugly mangled name like <myVar>5__1 was null.

Changes:
1. Formatter:
- Tweak implementation to handle compiler generated names.

2. CSharpFormatter:
- Add implementation for C#. VB is still TODO.
- Split up GeneratedNames so that it can be included in the CSharpResultProvider.
